requirements.txt
langchain==0.0.215
openai==0.27.2
python-dotenv==1.0.0
streamlit==1.22.0
